refactor: extract stepper, wizard, footer into own dirs
