+++ client/src/feat/my-passport-form/ui/wizard/types.ts.ts
@@ -1,0 +1,6 @@
+export interface WizardStep {
+	hash: string;
+	status: any;
+	label: string;
+	description: string;
+}
